fix(connection manager): unset ready on connection error
Currently, there is no way to reliably detect when the client re-connects, as the `Ready` event is only ever triggered on the first successful connection.

By unsetting the READY atomic here, the client will re-send an `Event::Ready`, allowing consumers to detect this.

Arguably this is a bit of a bodge-job, as it sort of negates the point for the ready check. It might be a better idea to instead add a separate event specific to (re)connecting, or just remove the check altogether.